requirements.txt
# Web framework
Flask==3.1.1
gunicorn==21.2.0

# Real-ESRGAN Stack
torch==2.0.1+cpu
torchvision==0.15.2+cpu
basicsr==1.4.2
realesrgan==0.3.0
facexlib==0.3.0
gfpgan==1.3.8

# Core Libraries
numpy==1.24.3
Pillow==10.0.1
opencv-python-headless==4.8.1.78
tqdm==4.66.1
